Move ui definitions to ui project
