+++ src/openrct2/interface/Viewport.cpp
@@ -76,6 +76,11 @@ InteractionInfo::InteractionInfo(const PaintStruct* ps)
 
 static void ViewportPaintWeatherGloom(DrawPixelInfo& dpi);
 static void ViewportPaint(const Viewport* viewport, DrawPixelInfo& dpi, const ScreenRect& screenRect);
+static void ViewportUpdateFollowSprite(WindowBase* window);
+static void ViewportUpdateSmartFollowEntity(WindowBase* window);
+static void ViewportUpdateSmartFollowStaff(WindowBase* window, const Staff* peep);
+static void ViewportUpdateSmartFollowVehicle(WindowBase* window);
+static void ViewportInvalidate(const Viewport* viewport, const ScreenRect& screenRect);
 
 /**
  * This is not a viewport function. It is used to setup many variables for
@@ -228,7 +233,7 @@ void ViewportRemove(Viewport* viewport)
     _viewports.erase(it);
 }
 
-Viewport* ViewportGetMain()
+static Viewport* ViewportGetMain()
 {
     auto mainWindow = WindowGetMain();
     if (mainWindow == nullptr)
